initial update on crud

+++ src/main/java/com/portfolio/diary/repository/UserInfo_Repository.java
@@ -5,7 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.portfolio.diary.model.UserInfo_Entity;
 
 
-
-public interface UserInfo_Repository extends JpaRepository<UserInfo_Entity,Long> {
-
+public interface UserInfo_Repository extends JpaRepository<UserInfo_Entity, Integer> {
 }
